fix: mypy type annotations for autonomy modules
Add type annotations to handler functions and private helpers
to satisfy mypy strict mode. Fix dict.get key function for max().

+++ ortobahn/adaptive_threshold.py
@@ -59,4 +59,4 @@ def compute_adaptive_threshold(
 
     avg_bias = sum(biases) / len(biases)
     adjusted = default + (avg_bias * ADJUSTMENT_RATE)
-    return max(MIN_THRESHOLD, min(MAX_THRESHOLD, round(adjusted, 3)))
+    return float(max(MIN_THRESHOLD, min(MAX_THRESHOLD, round(adjusted, 3))))
